+++ bin/check-es-cluster-status.rb
@@ -83,15 +83,10 @@ class ESClusterStatus < Sensu::Plugin::Check::CLI
          short: '-e',
          long: '--https'
 
-  option :yellow_only,
-         description: 'If you would only like to be alerted for a yellow (or green) state',
-         long: '--yellow_only',
-         boolean: true
-
-  option :red_only,
-         description: 'If you would only like to be alerted for a red (or green) state',
-         long: '--red_only',
-         boolean: true
+  option :alert_status,
+         description: 'Only alert when status matches given (red/yellow) option (or green)',
+         long: '--alert_status STATUS',
+         default: ''
 
   def get_es_resource(resource)
     headers = {}
@@ -149,13 +144,13 @@ class ESClusterStatus < Sensu::Plugin::Check::CLI
       when 'green'
         ok 'Cluster is green'
       when 'yellow'
-        if !options[:red_only]
+        if !options[:alert_status].include? 'red'
           warning 'Cluster state is Yellow'
         else
           ok 'Not alerting on yellow'
         end
       when 'red'
-        if !options[:yellow_only]
+        if !options[:alert_status].include? 'yellow'
           critical 'Cluster state is Red'
         else
           ok 'Not alerting on red'